refactor(xlsx): extract viewport render orchestration for worker reuse
